+++ src/Resolvers/NetteDatabase/HasManyCountResolver.php
@@ -2,13 +2,17 @@
 
 namespace Efabrica\GraphQL\Nette\Resolvers\NetteDatabase;
 
+use Efabrica\GraphQL\Exceptions\ResolverException;
 use Efabrica\GraphQL\Schema\Definition\ResolveInfo;
 use Nette\Database\Table\ActiveRow;
+use Throwable;
 
 final class HasManyCountResolver extends DatabaseResolver
 {
     /**
      * @param ActiveRow $parentValue
+     *
+     * @throws ResolverException
      */
     public function __invoke($parentValue, array $args, ResolveInfo $resolveInfo): int
     {
@@ -19,6 +23,10 @@ final class HasManyCountResolver extends DatabaseResolver
 
         $this->applyConditionsToSelection($selection, $args);
 
-        return $selection->count('*');
+        try {
+            return $selection->count('*');
+        } catch (Throwable $e) {
+            throw new ResolverException('There was an error while executing the query', 0, $e, $e);
+        }
     }
 }