+++ src/Resolvers/NetteDatabase/HasManyResolver.php
@@ -2,13 +2,17 @@
 
 namespace Efabrica\GraphQL\Nette\Resolvers\NetteDatabase;
 
+use Efabrica\GraphQL\Exceptions\ResolverException;
 use Efabrica\GraphQL\Schema\Definition\ResolveInfo;
 use Nette\Database\Table\ActiveRow;
+use Throwable;
 
 final class HasManyResolver extends DatabaseResolver
 {
     /**
      * @param ActiveRow $parentValue
+     *
+     * @throws ResolverException
      */
     public function __invoke($parentValue, array $args, ResolveInfo $resolveInfo): array
     {
@@ -21,6 +25,11 @@ final class HasManyResolver extends DatabaseResolver
         $this->applyOrderToSelection($selection, $args);
         $this->applyConditionsToSelection($selection, $args);
 
-        return $selection->fetchAll();
+
+        try {
+            return $selection->fetchAll();
+        } catch (Throwable $e) {
+            throw new ResolverException('There was an error while executing the query', 0, $e, $e);
+        }
     }
 }