+++ src/Resolvers/NetteDatabase/TableResolver.php
@@ -2,8 +2,10 @@
 
 namespace Efabrica\GraphQL\Nette\Resolvers\NetteDatabase;
 
+use Efabrica\GraphQL\Exceptions\ResolverException;
 use Efabrica\GraphQL\Schema\Definition\ResolveInfo;
 use Nette\Database\Table\ActiveRow;
+use Throwable;
 
 final class TableResolver extends DatabaseResolver
 {
@@ -11,6 +13,7 @@ final class TableResolver extends DatabaseResolver
      * @param null $parentValue
      *
      * @return ActiveRow[]
+     * @throws ResolverException
      */
     public function __invoke($parentValue, array $args, ResolveInfo $resolveInfo): array
     {
@@ -20,6 +23,10 @@ final class TableResolver extends DatabaseResolver
         $this->applyOrderToSelection($selection, $args);
         $this->applyConditionsToSelection($selection, $args);
 
-        return $selection->fetchAll();
+        try {
+            return $selection->fetchAll();
+        } catch (Throwable $e) {
+            throw new ResolverException('There was an error while executing the query', 0, $e, $e);
+        }
     }
 }