+++ CHANGELOG.md
@@ -2,10 +2,24 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2023-04-06
+### Added
+- Nette DI extension
+- AdditionalResponseData to NetteDatabaseResolverFactory and DatabaseResolver [BC]
+- Debug exceptions to resolvers 
+- Logging of SQL queries to addition debug data in DatabaseResolver
+
+### Changed
+- NetteDatabaseSchemaLoader accepts NetteDatabaseResolverFactoryInterface instead of NetteDatabaseResolverFactory [BC]
+
+### Removed
+- Example config.neon [BC]
+
 ## [0.1.0] - 2022-10-24
 ### Added
 - Nette database schema loader
 - Nette database resolvers
 
-[Unreleased]: https://github.com/efabrica-team/nette-graphql/compare/0.1.0...main
+[Unreleased]: https://github.com/efabrica-team/nette-graphql/compare/0.2.0...main
+[0.2.0]: https://github.com/efabrica-team/nette-graphql/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/efabrica-team/nette-graphql/compare/0.0.0...0.1.0